Base bounds getter in MeasuringTape.
See https://github.com/phetsims/charges-and-fields/issues/74 for the need. We don't want to have the user drag the measuring tape and have the long tape or tip overlap the toolbox (and then have the tape be returned to toolbox). Instead we care whether the user drags the base into the toolbox.